+++ src/main/java/io/fabric8/maven/docker/assembly/DockerAssemblyManager.java
@@ -1,10 +1,22 @@
 package io.fabric8.maven.docker.assembly;
 
-import java.io.*;
-import java.util.*;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.Writer;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
-import io.fabric8.maven.docker.config.*;
-import io.fabric8.maven.docker.util.*;
+import io.fabric8.maven.docker.config.ArchiveCompression;
+import io.fabric8.maven.docker.config.Arguments;
+import io.fabric8.maven.docker.config.AssemblyConfiguration;
+import io.fabric8.maven.docker.config.AssemblyMode;
+import io.fabric8.maven.docker.config.BuildImageConfiguration;
+import io.fabric8.maven.docker.util.DockerFileUtil;
+import io.fabric8.maven.docker.util.EnvUtil;
+import io.fabric8.maven.docker.util.Logger;
+import io.fabric8.maven.docker.util.MojoParameters;
 import org.apache.commons.io.IOUtils;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.model.Build;
@@ -29,7 +41,7 @@ import org.codehaus.plexus.archiver.util.DefaultArchivedFileSet;
 import org.codehaus.plexus.archiver.util.DefaultFileSet;
 import org.codehaus.plexus.component.annotations.Component;
 import org.codehaus.plexus.component.annotations.Requirement;
-
+import org.codehaus.plexus.interpolation.fixed.FixedStringSearchInterpolator;
 
 /**
  * Tool for creating a docker image tar ball including a Dockerfile for building
@@ -111,11 +123,9 @@ public class DockerAssemblyManager {
                                                      buildConfig.getDockerFile() + "\" (resolved to \"" + dockerFile + "\") doesn't exist");
                 }
 
-                verifyGivenDockerfile(dockerFile, buildConfig, params.getProject(), log);
-                Properties interpolationProperties = new Properties();
-                interpolationProperties.putAll(params.getProject().getProperties());
-                interpolationProperties.putAll(params.getSession().getSystemProperties());
-                interpolateDockerfile(dockerFile, buildDirs, interpolationProperties, buildConfig.getFilter());
+                FixedStringSearchInterpolator interpolator = DockerFileUtil.createInterpolator(params, buildConfig.getFilter());
+                verifyGivenDockerfile(dockerFile, buildConfig, interpolator, log);
+                interpolateDockerfile(dockerFile, buildDirs, interpolator);
                 // User dedicated Dockerfile from extra directory
                 customizer = new ArchiverCustomizer() {
                     @Override
@@ -168,21 +178,20 @@ public class DockerAssemblyManager {
         fileSet.setExcludes(excludes.toArray(new String[0]));
     }
 
-    private void interpolateDockerfile(File dockerFile, BuildDirs params, Properties properties, String filter) throws IOException {
+    private void interpolateDockerfile(File dockerFile, BuildDirs params, FixedStringSearchInterpolator interpolator) throws IOException {
         File targetDockerfile = new File(params.getOutputDirectory(), dockerFile.getName());
-        String dockerFileInterpolated =
-            DockerFileUtil.interpolate(dockerFile, properties, filter);
+        String dockerFileInterpolated = DockerFileUtil.interpolate(dockerFile, interpolator);
         try (Writer writer = new FileWriter(targetDockerfile)) {
             IOUtils.write(dockerFileInterpolated, writer);
         }
     }
 
-    private void verifyGivenDockerfile(File dockerFile, BuildImageConfiguration buildConfig, MavenProject project, Logger log) throws IOException {
+    private void verifyGivenDockerfile(File dockerFile, BuildImageConfiguration buildConfig, FixedStringSearchInterpolator interpolator, Logger log) throws IOException {
         AssemblyConfiguration assemblyConfig = buildConfig.getAssemblyConfiguration();
         if (assemblyConfig != null) {
             String name = assemblyConfig.getName();
             for (String keyword : new String[] { "ADD", "COPY" }) {
-                List<String[]> lines = DockerFileUtil.extractLines(dockerFile, keyword, project.getProperties(), buildConfig.getFilter());
+                List<String[]> lines = DockerFileUtil.extractLines(dockerFile, keyword, interpolator);
                 for (String[] line : lines) {
                     // contains an ADD/COPY ... targetDir .... All good.
                     if (!line[0].startsWith("#") && line.length > 1 && line[1].contains(name)) {