+++ src/main/java/io/fabric8/maven/docker/service/BuildService.java
@@ -230,8 +230,7 @@ public class BuildService {
             File fullDockerFilePath = buildConfig.getAbsoluteDockerFilePath(buildContext.getMojoParameters());
             fromImage = DockerFileUtil.extractBaseImage(
                 fullDockerFilePath,
-                buildContext.getMojoParameters().getProject().getProperties(),
-                buildConfig.getFilter());
+                DockerFileUtil.createInterpolator(buildContext.getMojoParameters(), buildConfig.getFilter()));
         } catch (IOException e) {
             // Cant extract base image, so we wont try an auto pull. An error will occur later anyway when
             // building the image, so we are passive here.